pyproject.toml
[tool.poetry]
name = "aidev3"
version = "0.1.0"
description = ""
authors = ["Jakub Rosa <jakub.rosa.95@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
requests-random-user-agent = "^2023.2.15"
beautifulsoup4 = "^4.12.3"
langfuse = "^2.53.9"
openai = "^1.54.3"
tinytag = "^2.0.0"
pillow = "^11.0.0"
pytesseract = "^0.3.13"
tqdm = "^4.67.0"
pyhtml2md = "^1.6.0"
qdrant-client = "^1.12.1"
neo4j = "^5.26.0"
flask = "^3.1.0"
pymupdf = "^1.24.14"
opencv-python = "^4.10.0.84"
litellm = "^1.53.1"
thefuzz = "^0.22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
